# backend/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: NNIKITI API
  description: This is an API for the educational institution's information system.
servers:
  - url: http://127.0.0.1:8000/api/v1/
    description: Local development
...
paths:
...
  /auth/token/:
    post:
      tags:
        - auth
      summary: Refresh access token using refresh_token cookie
      description: >
        1) Send POST with body:
           - `client_id=client_id`
           - `grant_type=refresh_token`
        
        2) **Refresh token is NOT sent in the body; it is automatically read from the HttpOnly cookie `refresh_token`.**
        
        3) If the refresh token is valid, a new access token is returned in the response body.
        
        4) A new refresh token cookie is also set to replace the old one.
        *This endpoint does NOT require an Authorization header because refresh token is taken from the cookie.*
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [ client_id ,grant_type ]
              properties:
                client_id:
                  type: string
                  enum: [client_id]
                  description: Client token from API
                grant_type:
                  type: string
                  enum: [ refresh_token ]
                  description: Must be exactly `refresh_token`
      responses:
        "200":
          description: Access token successfully refreshed
          headers:
            Set-Cookie:
              description: Updated refresh token set in cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: New access token
                  expires_in:
                    type: integer
                    description: Access token lifetime in seconds
                  token_type:
                    type: string
                    example: Bearer
        "400":
          description: Invalid request or missing/invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    AccessTokenAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Human readable error message
      required:
        - code
        - message
...